requirements.txt
fastapi>=0.110,<1.0
uvicorn[standard]>=0.29,<1.0
python-multipart>=0.0.9
pillow>=10.0
numpy>=1.24
scikit-image>=0.21
tensorflow>=2.12
